wip quest system

+++ gbCore/sources/scene_graph.h
@@ -34,7 +34,7 @@ namespace gb
         camera_2d_shared_ptr m_camera_2d;
         camera_3d_shared_ptr m_camera_3d;
         
-        glm::vec2 m_scene_size;
+        glm::ivec2 m_scene_size;
         
     public:
